+++ features/step_definitions/idv_steps/idv_steps.rb
@@ -21,7 +21,7 @@ end
 
 Given(/^I should see the error message for identity verification "(.*?)"$/) do |expected_text|
   sleep 15
-  (idv_error_message.should eq(expected_text)).should == true
+  idv_error_message.should eq(expected_text)
 end
 
 Given("I {string} via phone using a mocked driver's license") do |action|